requirements.txt
IPy==1.01
qqwry-py3==1.1.0